changed client to front-end to fix 404 error on npm run dev
