--- client/package.json
@@ -1,29 +1,0 @@
-{
-  "name": "client",
-  "version": "0.0.0",
-  "main": "index.js",
-  "scripts": {
-    "dev": "vite",
-    "build": "vite build",
-    "serve": "vite preview"
-  },
-  "dependencies": {
-      "@apollo/client": "^3.10.4",
-      "@chakra-ui/react": "^2.8.2",
-      "@emotion/react": "^11.11.4",
-      "@emotion/styled": "^11.11.5",
-      "axios": "^1.7.2",
-      "cors": "^2.8.5",
-      "framer-motion": "^11.2.9",
-      "graphql": "^16.8.1",
-      "jwt-decode": "^4.0.0",
-      "nodemon": "^3.1.3",
-      "react": "^18.3.1",
-      "react-dom": "^18.3.1",
-      "react-router-dom": "^6.23.1"
-  },
-  "devDependencies": {
-    "vite": "^4.2.0",
-    "@vitejs/plugin-react": "^3.0.0"
-  }
-}
