+++ front-end/src/App.jsx
@@ -1,0 +1,23 @@
+import React from "react";
+// import { BrowserRouter as Router, Routes, Route } from 'react-router-dom';
+// import LoginPage from './pages/LoginPage';
+// import SignupPage from './pages/SignupPage';
+// import DashboardPage from './pages/DashboardPage';
+
+function App() {
+  return (
+    <div>
+      <h1>hello</h1>
+    </div>
+  );
+}
+
+export default App;
+
+// <Router>
+//   <Routes>
+//     <Route path="/" element={<LoginPage />} />
+//     <Route path="/signup" element={<SignupPage />} />
+//     <Route path="/dashboard" element={<DashboardPage />} />
+//   </Routes>
+// </Router>
