+++ front-end/src/GlobalStyles.jsx
