+++ front-end/src/components/SpendingHistory.jsx
